+++ client_server/client/CMakeLists.txt
@@ -3,8 +3,9 @@ project(Client_Server VERSION 1.0)
 
 project(Client_Server)
 
-set(SUBDIRS ./../clientConfigParserHelper ./../clientServerParserLib ./../maxValues)
+set(SUBDIRS ./../clientConfigParserHelper ./../clientServerParserLib ./../maxValues ./../interacting_with_process)
 
-include_directories(./../cxxopts  ./../json.h ${SUBDIRS})
+include_directories(./../cxxopts ./../json.h ${SUBDIRS})
 
-add_executable(client client.cpp ./../maxValues/maxAmpsVoltsParser.cpp ./../clientServerParserLib/clientServerParserLib.cpp ./../clientConfigParserHelper/clientConfigParser.cpp)
+add_executable(client client.cpp ./../maxValues/maxAmpsVoltsParser.cpp ./../clientServerParserLib/clientServerParserLib.cpp
+        ./../clientConfigParserHelper/clientConfigParser.cpp ./../interacting_with_process/interacting_with_process.cpp)
