+++ client_server/server/server.cpp
@@ -13,59 +13,6 @@
 // limitations under the License.
 // =============================================================================
 
-#undef UNICODE
-
-#define WIN32_LEAN_AND_MEAN
-
-#include <windows.h>
-#include <winsock2.h>
-#include <ws2tcpip.h>
-#include <stdlib.h>
-#include <stdio.h>
-#include <tlhelp32.h>
-#include <tchar.h>
-#include <iostream>
-#include <processthreadsapi.h>
-#include <stdlib.h>
-#include <algorithm>
-#include <fstream>
-#include <netinet/tcp.h>
-#include "./../cxxopts/include/cxxopts.hpp"
-#include "./../serverConfigParserHelper/serverConfigParser.h"
-#include <errno.h>
-#include <stdio.h>
-#include <fcntl.h>
-#include <sys/stat.h>
-
-#pragma comment (lib, "Ws2_32.lib")
-
-#define START_PTD 100
-#define STOP_PTD 200
-#define GET_DATA 500
-
-#define PTD_PORT "8888"
-#define PTD_IP "127.0.0.1"
-#define MINUTE_DURATION_IN_SECONDS 60
-
-#define DEFAULT_BUFFER_CHUNK_SIZE 4096
-#define DEFAULT_FILE_CHUNK_SIZE 65536
-#define DEFAULT_BUFLEN 512
-
-struct serverAnswer {
-    int code;
-    char msg[DEFAULT_BUFLEN];
-};
-
-struct MaxAmpsVolts {
-    float maxAmps;
-    float maxVolts;
-};
-
-struct InitMessage {
-    int messageNumber;
-    MaxAmpsVolts maxValues;
-};
-
 int sentMessage(int ClientSocket, char *message, size_t msgLength) {
     int iSendResult = send(ClientSocket, message, msgLength, 0);
     if (iSendResult == SOCKET_ERROR) {
@@ -90,105 +37,6 @@ int sentAnswer(int ClientSocket, int code, char *msg) {
     answer.code = code;
     sprintf(answer.msg, msg);
     return sentAnswerForClient(ClientSocket, &answer);
-}
-
-int64_t GetFileSize(const std::string &fileName) {
-    FILE *f;
-    if ((f = fopen(fileName.c_str(), "rb")) == NULL) {
-        return -1;
-    }
-    fseek(f, 0, SEEK_END);
-    const int64_t len = ftell(f);
-    fclose(f);
-    return len;
-}
-
-int SendBuffer(SOCKET s, const char *buffer, int bufferSize, int chunkSize = DEFAULT_BUFFER_CHUNK_SIZE) {
-    int allSendedBytes = 0;
-    while (allSendedBytes < bufferSize) {
-        const int sendedBytes = send(s, &buffer[allSendedBytes], std::min(chunkSize, bufferSize - allSendedBytes), 0);
-        if (sendedBytes < 0) {
-            return sendedBytes;
-        }
-        allSendedBytes += sendedBytes;
-    }
-    return allSendedBytes;
-}
-
-int64_t SendFile(SOCKET s, const std::string &fileName, int chunkSize = DEFAULT_FILE_CHUNK_SIZE) {
-    const int64_t fileSize = GetFileSize(fileName);
-
-    if (fileSize < 0) {
-        std::cerr << "Can not get file size" << std::endl;
-        return 1;
-    }
-
-    std::ifstream file(fileName, std::ifstream::binary);
-    if (file.fail()) {
-        std::cerr << "Can not open ifstream" << std::endl;
-        return 1;
-    }
-
-    // Send file size
-    if (SendBuffer(s, reinterpret_cast<const char *>(&fileSize),
-                   sizeof(fileSize)) != sizeof(fileSize)) {
-        std::cerr << "Can not send file size" << std::endl;
-        return 1;
-    }
-
-    char *buffer = new char[chunkSize];
-    bool errored = false;
-    int64_t i = fileSize;
-    while (i != 0) {
-        const int64_t ssize = std::min(i, (int64_t) chunkSize);
-        if (!file.read(buffer, ssize)) {
-            errored = true;
-            break;
-        }
-        const int l = SendBuffer(s, buffer, (int) ssize);
-        if (l < 0) {
-            errored = true;
-            break;
-        }
-        i -= l;
-    }
-    delete[] buffer;
-    if (!errored) {
-        std::cout << "Send file to client" << std::endl;
-    } else {
-        std::cerr << "Can not send file to client" << std::endl;
-    }
-
-    file.close();
-
-    return errored ? 1 : fileSize;
-}
-
-bool executeSystemCommand(char *command_Line, STARTUPINFO *si, PROCESS_INFORMATION *pi) {
-    ZeroMemory(si, sizeof(*si));
-    si->cb = sizeof(si);
-    ZeroMemory(pi, sizeof(*pi));
-
-    // Start the child process.
-    if (!CreateProcess(NULL,   // No module name (use command line)
-                       command_Line,        // Command line
-                       NULL,           // Process handle not inheritable
-                       NULL,           // Thread handle not inheritable
-                       FALSE,          // Set handle inheritance to FALSE
-                       CREATE_NEW_CONSOLE, // No creation flags
-                       NULL,           // Use parent's environment block
-                       NULL,           // Use parent's starting directory
-                       si,            // Pointer to STARTUPINFO structure
-                       pi)           // Pointer to PROCESS_INFORMATION structure
-            ) {
-        std::cerr << "CreateProcess failed " << GetLastError() << "Exit (1)" << std::endl;
-        return (false);
-    }
-    return (true);
-}
-
-bool closeSystemProcess(PROCESS_INFORMATION *pi) {
-    return TerminateProcess(pi->hProcess, 0) && CloseHandle(pi->hProcess) && CloseHandle(pi->hThread);
 }
 
 void recvPtdAnswer(int ptdClientSocket) {
